refactor: 테스트 클래스 패키지 구조를 프로덕션 코드와 일치하도록 변경
